# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface ModelInfo {
  /** One of zhi1, zhi2, zhi3, zhi4 */
  id: string;
  /** Display label (e.g. "Zhi 1") */
  label: string;
  /** Underlying provider name */
  provider: string;
  /** Whether this model is safe for explicit/X-rated text */
  supportsExplicit: boolean;
}

export interface ExtractedText {
  text: string;
  filename: string;
  characters?: number;
}

export interface ReferenceImage {
  /** e.g. "the man" or "Sarah" */
  label: string;
  /** data:image/...;base64 encoded reference image */
  dataUrl: string;
}

export interface NovelInput {
  title?: string;
  /** @minLength 1 */
  sourceText: string;
  /**
     * @minimum 1
     * @maximum 50
     */
  panelCount: number;
  /** User's free-form direction (style, characters, tone) */
  specifications: string;
  /** zhi1 | zhi2 | zhi3 | zhi4 */
  textModel: string;
  /** Force the X-rated path through Venice for text */
  explicit?: boolean;
  /** e.g. "noir comic", "manga", "watercolor" */
  artStyle?: string;
  referenceImages?: ReferenceImage[];
}

export interface NovelSummary {
  id: number;
  title: string;
  /** pending | generating | done | failed */
  status: string;
  panelCount: number;
  completedPanels: number;
  createdAt: string;
  /** @nullable */
  coverImage?: string | null;
  textModel?: string;
  /** @nullable */
  artStyle?: string | null;
}

export interface Panel {
  id: number;
  idx: number;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
  /** @nullable */
  imageDataUrl?: string | null;
  /** pending | generating | done | failed */
  status: string;
  /** @nullable */
  error?: string | null;
}

export interface Novel {
  id: number;
  title: string;
  sourceText: string;
  specifications: string;
  panelCount: number;
  textModel: string;
  /** @nullable */
  artStyle?: string | null;
  explicit?: boolean;
  status: string;
  /** @nullable */
  error?: string | null;
  createdAt: string;
  panels: Panel[];
}

export interface ScreenplayInput {
  title?: string;
  /** @minLength 1 */
  sourceText: string;
  /** How to handle the conversion (non-fiction expository, framed in a story, etc.) */
  specifications: string;
  /** zhi1 | zhi2 | zhi3 | zhi4 */
  textModel: string;
  explicit?: boolean;
}

export interface ScreenplaySummary {
  id: number;
  title: string;
  textModel: string;
  createdAt: string;
}

export interface Screenplay {
  id: number;
  title: string;
  sourceText: string;
  specifications: string;
  textModel: string;
  explicit?: boolean;
  content: string;
  createdAt: string;
}