# lib/api-zod/src/generated/types/extractedText.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedText {
  text: string;
  filename: string;
  characters?: number;
}