# lib/api-zod/src/generated/types/modelInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelInfo {
  /** One of zhi1, zhi2, zhi3, zhi4 */
  id: string;
  /** Display label (e.g. "Zhi 1") */
  label: string;
  /** Underlying provider name */
  provider: string;
  /** Whether this model is safe for explicit/X-rated text */
  supportsExplicit: boolean;
}