# lib/api-zod/src/generated/types/novel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { Panel } from './panel';

export interface Novel {
  id: number;
  title: string;
  sourceText: string;
  specifications: string;
  panelCount: number;
  textModel: string;
  /** @nullable */
  artStyle?: string | null;
  explicit?: boolean;
  status: string;
  /** @nullable */
  error?: string | null;
  createdAt: string;
  panels: Panel[];
}