# lib/api-zod/src/generated/types/novelInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ReferenceImage } from './referenceImage';

export interface NovelInput {
  title?: string;
  /** @minLength 1 */
  sourceText: string;
  /**
     * @minimum 1
     * @maximum 50
     */
  panelCount: number;
  /** User's free-form direction (style, characters, tone) */
  specifications: string;
  /** zhi1 | zhi2 | zhi3 | zhi4 */
  textModel: string;
  /** Force the X-rated path through Venice for text */
  explicit?: boolean;
  /** e.g. "noir comic", "manga", "watercolor" */
  artStyle?: string;
  referenceImages?: ReferenceImage[];
}