# lib/api-zod/src/generated/types/novelSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface NovelSummary {
  id: number;
  title: string;
  /** pending | generating | done | failed */
  status: string;
  panelCount: number;
  completedPanels: number;
  createdAt: string;
  /** @nullable */
  coverImage?: string | null;
  textModel?: string;
  /** @nullable */
  artStyle?: string | null;
}