# lib/api-zod/src/generated/types/panel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Panel {
  id: number;
  idx: number;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
  /** @nullable */
  imageDataUrl?: string | null;
  /** pending | generating | done | failed */
  status: string;
  /** @nullable */
  error?: string | null;
}