# lib/api-zod/src/generated/types/referenceImage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface ReferenceImage {
  /** e.g. "the man" or "Sarah" */
  label: string;
  /** data:image/...;base64 encoded reference image */
  dataUrl: string;
}