# lib/api-zod/src/generated/types/screenplay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Screenplay {
  id: number;
  title: string;
  sourceText: string;
  specifications: string;
  textModel: string;
  explicit?: boolean;
  content: string;
  createdAt: string;
}