# lib/api-zod/src/generated/types/screenplayInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface ScreenplayInput {
  title?: string;
  /** @minLength 1 */
  sourceText: string;
  /** How to handle the conversion (non-fiction expository, framed in a story, etc.) */
  specifications: string;
  /** zhi1 | zhi2 | zhi3 | zhi4 */
  textModel: string;
  explicit?: boolean;
}