# lib/api-zod/src/generated/types/screenplaySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Graphic Novel Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface ScreenplaySummary {
  id: number;
  title: string;
  textModel: string;
  createdAt: string;
}